PCHR-2647: Split base listener in multiple event based listeners
