+++ uk.co.compucorp.civicrm.hrcore/CRM/HRCore/HookListener/ObjectBased/Page/CaseDashboard.php
@@ -1,6 +1,6 @@
 <?php
 
-class CRM_HRCore_HookListener_Page_CaseDashboard extends CRM_HRCore_HookListener_ObjectBasedListener {
+class CRM_HRCore_HookListener_ObjectBased_Page_CaseDashboard extends CRM_HRCore_HookListener_ObjectBased_ObjectBasedListener {
 
   protected $objectClass = 'CRM_Case_Page_DashBoard';
   protected $pageCSSIdentifier = '.page-civicrm-case';
