+++ uk.co.compucorp.civicrm.hrcore/CRM/HRCore/HookListener/ObjectBased/Page/Dashlet/CaseDashboard.php
@@ -1,0 +1,6 @@
+<?php
+
+class CRM_HRCore_HookListener_ObjectBased_Page_Dashlet_CaseDashboard extends CRM_HRCore_HookListener_ObjectBased_Page_CaseDashboard {
+
+  protected $pageCSSIdentifier = '#case_dashboard_dashlet';
+}
