--- uk.co.compucorp.civicrm.hrcore/CRM/HRCore/HookListener/Page/Dashlet/CaseDashboard.php
@@ -1,6 +1,0 @@
-<?php
-
-class CRM_HRCore_HookListener_Page_Dashlet_CaseDashboard extends CRM_HRCore_HookListener_Page_CaseDashboard {
-
-  protected $pageCSSIdentifier = '#case_dashboard_dashlet';
-}
